+++ packages/amplify-category-api/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.15.1](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/amplify-category-api@5.15.0...@aws-amplify/amplify-category-api@5.15.1) (2025-09-25)
+
+**Note:** Version bump only for package @aws-amplify/amplify-category-api
+
 # [5.15.0](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/amplify-category-api@5.14.2...@aws-amplify/amplify-category-api@5.15.0) (2025-04-03)
 
 ### Features